feat(locations): add frontend Location types (LOC-02)
Add dedicated frontend/src/types/location.ts with Location,
CreateLocationInput, UpdateLocationInput, and BulkImportResult to mirror
the backend aggregate from LOC-01. Move the duplicate Location interface
out of event.ts and re-export it for backward compatibility.
